[MCUX-57045] Remove mbedtls as a required field from port layer
Port layer doesn't require the compomnent mbedtls. It should be
added in final example. This will also give ability to user to select
whether they want to use all mbedtls files, only crypto or only x509.

This would be needed for resource constrained systems and even when
using the port layer with TF-M.

Signed-off-by: Ruchika Gupta <ruchika.gupta_1@nxp.com>