example: Add example to illustrate generation and usage of keys.
The example illustrates the generation and usage of following keys:
1. ECC keys
2. AES keys
3. RSA keys
4. MAC Keys

The location of these keys can be chosen in the yml file for the boards
which the example is being included for. By deafult even though the
example name says opaque, the default location in test is LOCAL_STORAGE.
This is chosen as default as it would work for all platforms.

However the intent is that this example can be built to test in opaque
location keys. For this, the calling yml needs to define the location of
TEST KEYS i.e TEST_KEY_LOCATION.
eg. For RT1180, we choose this location as PSA_CRYPTO_ELE_S4XX_LOCATION

Signed-off-by: Ruchika Gupta <ruchika.gupta_1@nxp.com>